feat(api): move and rename function acceptPixLastTermsOfService
from IAM to acceptLegacyPixAppTermsOfService in legal-documents userRepository
